Safari-609.1.11.1/LayoutTests/http/tests/misc/location-replace-crossdomain-async-delegates-expected.txt

Checks that a cross origin iframe can be navigated using location.replace().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Navigated cross-origin frame via Location.replace
PASS window.frames['crossDomainFrame'].location.href is "http://127.0.0.1:8000/security/resources/postMessage.html"
PASS successfullyParsed is true

TEST COMPLETE
